fix(PreProcessing): delete useless concept and modify place of it
