+++ include/Engine/Container/TriggerMap.hpp
@@ -6,84 +6,7 @@
 #include <cstddef>
 
 #include "Engine/System/ThreadPool.hpp"
-
-template<class T, class ...Ts>
-struct tuple_contain;
-
-template<class T, class _T, class ...Ts>
-struct tuple_contain<T, _T, Ts...>
-{
-    static constexpr bool value = std::is_same<T, _T>::value || tuple_contain<T, Ts...>::value;
-};
-
-template<class T, class _T>
-struct tuple_contain<T, _T>
-{
-    static constexpr bool value = std::is_same<T, _T>::value;
-};
-
-namespace imp
-{
-    template<size_t I, class T, class ...Ts>
-    struct tuple_find;
-
-    template<size_t I, class T, class _T, class ...Ts>
-    struct tuple_find<I, T, _T, Ts...>
-    {
-        static constexpr size_t value = std::is_same<T, _T>::value ? I : imp::tuple_find<I, T, Ts...>::value;
-    };
-
-    template<size_t I, class T, class _T>
-    struct tuple_find<I, T, _T>
-    {
-        static constexpr size_t value = std::is_same<T, _T>::value ? I : I + 1;
-    };
-}
-
-template<class T, class ...Ts>
-struct tuple_find
-{
-    static constexpr size_t value = imp::tuple_find<0, T, Ts...>::value;
-};
-
-template<class ...Ts>
-struct tuple_size;
-
-template<class T, class ...Ts>
-struct tuple_size<T, Ts...>
-{
-    static constexpr size_t value = 1 + tuple_size<Ts...>::value;
-};
-
-template<class T>
-struct tuple_size<T>
-{
-    static constexpr size_t value = 1;
-};
-
-template<class, class>
-struct tuple_append;
-
-template<class T, class ...Ts>
-struct tuple_append<T, std::tuple<Ts...>>
-{
-    using type = std::tuple<T, Ts...>;
-};
-
-template<template<class> class T, class ...Ts>
-struct tuple_apply_template;
-
-template<template<class> class T, class _T, class ...Ts>
-struct tuple_apply_template<T, _T, Ts...>
-{
-    using type = typename tuple_append<T<_T>, typename tuple_apply_template<T, Ts...>::type>::type;
-};
-
-template<template<class> class T, class _T>
-struct tuple_apply_template<T, _T>
-{
-    using type = std::tuple<T<_T>>;
-};
+#include "Tool/PreProcessing.hpp"
 
 namespace eng
 {
@@ -109,7 +32,9 @@ namespace eng
             [[nodiscard]] constexpr T<_T> &at() noexcept;
 
         private:
-            typename tuple_apply_template<T, Ts...>::type m_tup;
+            //tuple_apply_template<T, Ts...>::type m_tup;
+            std::tuple<T<Ts>...> m_tup;
     };
 }
-#include "TriggerMap.inl"
+
+#include "Engine/Container/TriggerMap.inl"
