+++ include/Engine/Event/Trigger.inl
@@ -3,13 +3,13 @@
 
 namespace eng
 {
-    template<IsIEvent T>
+    template<class T>
     Trigger<T>::Trigger(ThreadPool& _tp)
         : m_tp(_tp)
     {
     }
 
-    template<IsIEvent T>
+    template<class T>
     void Trigger<T>::raise(const T &_event)
     {
         for (auto &_sub : m_sub)
@@ -18,14 +18,14 @@ namespace eng
             });
     }
 
-    template<IsIEvent T>
+    template<class T>
     Trigger<T>::sTask Trigger<T>::subscribe(Task _task)
     {
         m_sub.push_back(std::make_shared<Task>(_task));
         return m_sub.back();
     }
 
-    template<IsIEvent T>
+    template<class T>
     void Trigger<T>::unsubscribe(const sTask _task)
     {
         std::erase_if(m_sub, [_task] (const std::shared_ptr<Task> _ctask) {
